requirements.txt
# 深度学习框架
torch>=1.7.0
torchvision>=0.8.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0

# 数据处理
numpy>=1.19.0
scikit-learn>=0.24.0

# 可视化
matplotlib>=3.3.0

# GUI框架
PyQt5>=5.15.0

# 其他依赖
scipy>=1.5.0
